removed the dictionary structure that had all pairwise combinations  because the assumptions made in the program mean I don't have to worry about pair ids being switched.
