+++ relatednessFinder/database/__init__.py
@@ -1,1 +1,1 @@
-from .database_methods import get_connection
+from .database_methods import get_connection, get_relatedness
